0.4. generators + some menu stuff

+++ src/com/company/Vehicles/Van.java
@@ -1,10 +1,10 @@
 package com.company.Vehicles;
 
 public class Van extends Car{
-    public Integer loadArea;
+    public Double loadArea;
 
     public Van(Double value, String mark, String model, Integer mileage,
-               String color, String segment, Part[] parts, Integer loadArea) {
+               String color, String segment, Part[] parts, Double loadArea) {
         super(value, mark, model, mileage, color, segment, parts);
         this.loadArea = loadArea;
     }
